update terbaru ok

+++ donjo-app/models/Database_model.php
@@ -24,7 +24,9 @@
 		'19.02' => array('migrate' => 'nop', 'nextVersion' => '19.03'),
 		'19.03' => array('migrate' => 'migrasi_1903_ke_1904', 'nextVersion' => '19.04'),
 		'19.04' => array('migrate' => 'migrasi_1904_ke_1905', 'nextVersion' => '19.05'),
-		'19.05' => array('migrate' => 'migrasi_1905_ke_1906', 'nextVersion' => NULL)
+		'19.05' => array('migrate' => 'migrasi_1905_ke_1906', 'nextVersion' => '19.06'),
+		'19.06' => array('migrate' => 'migrasi_1906_ke_1907', 'nextVersion' => '19.07'),
+		'19.07' => array('migrate' => 'migrasi_1907_ke_1908', 'nextVersion' => NULL)
 	);
 
 	public function __construct()
@@ -181,36 +183,30 @@
 	$this->migrasi_1903_ke_1904();
 	$this->migrasi_1904_ke_1905();
 	$this->migrasi_1905_ke_1906();
+	$this->migrasi_1906_ke_1907();
+	$this->migrasi_1907_ke_1908();
+  }
+
+  private function migrasi_1907_ke_1908()
+  {
+  	$this->load->model('migrations/migrasi_1907_ke_1908');
+  	$this->migrasi_1907_ke_1908->up();
+  }
+
+  private function migrasi_1906_ke_1907()
+  {
+  	$this->load->model('migrations/migrasi_1906_ke_1907');
+  	$this->migrasi_1906_ke_1907->up();
   }
 
   private function migrasi_1905_ke_1906()
   {
-  	// Tambah tautan pada teks berjalan
-  	if (!$this->db->field_exists('tautan', 'teks_berjalan'))
-  	{
-			// Tambah kolom
-			$fields = array();
-			$fields['tautan'] = array(
-					'type' => 'varchar',
-					'constraint' => 150,
-			);
-			$fields['judul_tautan'] = array(
-					'type' => 'varchar',
-					'constraint' => 150,
-			);
-			$this->dbforge->add_column('teks_berjalan', $fields);
-		}
-
   	// Tambah kolom waktu update dan user pengupdate
   	if (!$this->db->field_exists('created_at', 'tweb_penduduk'))
   	{
 			// Tambah kolom
+			$this->dbforge->add_field("created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP");
 			$fields = array();
-			$fields['created_at'] = array(
-					'type' => 'timestamp',
-					'default CURRENT_TIMESTAMP',
-				  'null' => FALSE,
-			);
 			$fields['created_by'] = array(
 					'type' => 'int',
 					'constraint' => 11,
@@ -221,18 +217,25 @@
   	if (!$this->db->field_exists('updated_at', 'tweb_penduduk'))
   	{
 			// Tambah kolom
+			$this->dbforge->add_field("updated_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP");
 			$fields = array();
-			$fields['updated_at'] = array(
-					'type' => 'timestamp',
-					'default CURRENT_TIMESTAMP',
-				  'null' => FALSE,
-			);
 			$fields['updated_by'] = array(
 					'type' => 'int',
 					'constraint' => 11,
 				  'null' => FALSE,
 			);
 			$this->dbforge->add_column('tweb_penduduk', $fields);
+		}
+		else
+		{
+			$fields = array();
+			$fields['updated_by'] = array(
+					'type' => 'INT',
+					'constraint' => 11,
+				  'null' => TRUE,
+					'default' => NULL
+			);
+		  $this->dbforge->modify_column('tweb_penduduk', $fields);
 		}
 
   	// Tambah menu teks berjalan
@@ -310,6 +313,21 @@
 				}
 			}
 		}
+  	// Tambah tautan pada teks berjalan
+  	if (!$this->db->field_exists('tautan', 'teks_berjalan'))
+  	{
+			// Tambah kolom
+			$fields = array();
+			$fields['tautan'] = array(
+					'type' => 'varchar',
+					'constraint' => 150,
+			);
+			$fields['judul_tautan'] = array(
+					'type' => 'varchar',
+					'constraint' => 150,
+			);
+			$this->dbforge->add_column('teks_berjalan', $fields);
+		}
 
   	// Hapus menu SID dan Donasi
 		$this->db->where('id', 16)->delete('setting_modul');
@@ -337,6 +355,52 @@
   	}
   	// Di tweb_penduduk ubah kelahiran_anak_ke supaya default NULL
 	  $this->dbforge->modify_column('tweb_penduduk', array('kelahiran_anak_ke' => array('type' => 'TINYINT', 'constraint' => 2, 'default' => NULL)));
+
+	  // Ubah kolom tweb_penduduk supaya boleh null
+		$fields = array();
+		$fields['ktp_el'] = array(
+				'type' => 'TINYINT',
+				'constraint' => 4,
+			  'null' => TRUE,
+				'default' => NULL
+		);
+		$fields['status_rekam'] = array(
+				'type' => 'TINYINT',
+				'constraint' => 4,
+			  'null' => TRUE,
+				'default' => NULL
+		);
+		$fields['tempat_dilahirkan'] = array(
+				'type' => 'TINYINT',
+				'constraint' => 2,
+			  'null' => TRUE,
+				'default' => NULL
+		);
+		$fields['jenis_kelahiran'] = array(
+				'type' => 'TINYINT',
+				'constraint' => 2,
+			  'null' => TRUE,
+				'default' => NULL
+		);
+		$fields['penolong_kelahiran'] = array(
+				'type' => 'TINYINT',
+				'constraint' => 2,
+			  'null' => TRUE,
+				'default' => NULL
+		);
+		$fields['panjang_lahir'] = array(
+				'type' => 'VARCHAR',
+				'constraint' => 10,
+			  'null' => TRUE,
+				'default' => NULL
+		);
+		$fields['sakit_menahun_id'] = array(
+				'type' => 'INT',
+				'constraint' => 11,
+			  'null' => TRUE,
+				'default' => NULL
+		);
+	  $this->dbforge->modify_column('tweb_penduduk', $fields);
   }
 
   private function migrasi_1904_ke_1905()
